backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
Pillow==8.0.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
